Replace blank NSString with default '???'
It’s possible for the dataSource to return an empty string. If this
occurs the drawing code will crash as the size will be CGSizeZero.
Therefore replace empty stings with a default value.
If the user wishes to indicate a blank value they can do so by
returning a tab or ellipsis, or whatever else from the dataSource
method.